Introduce an adaptive mesh probe algorithm
At the high level, probe twice, if they are close enough together
continue to the next point.  If they aren't retry a configurable
number of times, throwing out the worst of the three points until they
are close enough or it is time to give up.
This allows both identifying bad Z probes (the assumption is a bad
probe will not be at a consistent height), and only retry those
locations as opposed to say take 3 readings throw out one, which means
it always takes longer, and if 2/3 are bad it has the wrong answer.