Z home then adaptive probing
The adaptive probing assumes the bed is close and doesn't work for the
initial Z probe.  This is to probe once to establish zero, then do an
adaptive probe to probe at least twice and make sure those are
consistent for a total of three probes.  The ideal would be two, but
the probe_at_point is written to assume Z has been homed and will
abort if it travels very far.