G30 disable probe sanity_check
Using G30 to measure the height of a model is very useful when a print
fails.  By default sanity_check is enabled, which in my case means I
couldn't measure anything higher than about 6mm.  The value is return
to the user, no need to prevent giving it to them.