Do adaptive probe right after one Z home without bump
This is to probe once to establish zero, then do an adaptive probe to
probe at least twice and make sure those are consistent for a total of
three probes.  The idea would be two, but the probe_at_point is
written to assume Z has been homed and will abort if it travels very
far when that isn't the case.