Add function comments
Add header comment for wait_for_confirmation so callers are aware that
it starts a timer that will shutoff and disable the hot end, that will
need to be stopped.  If it expires and shuts off, setting a higher
temperature from the LCD or host will fail to heat and throw a
firmware error requiring a reset.

The nozzle_timeout constant was defined earlier in the function and
has the same value, better to only have one.

set_probing_paused doesn't pause probing, but pauses other motors for
probing, adding a comment.

Explain why const_float_t as the previous comment still had me
thinking why bother?  It's a memory trade off, 2 bytes less, though
now with a pointer (reference), it will have to read the pointer
to get at the value.