Comment do_z_clearance, wait_for_confirmation remove redundant
Add header comment for wait_for_confirmation so callers are aware that
it starts a timer that will shutoff and disable the hot end, that will
need to be stopped.  If it expires and shuts off, setting a higher
temperature from the LCD or host will fail to heat and throw a
firmware error requiring a reset.

The nozzle_timeout constant is already available and has the same
value, easier to maintain when there's only one.