Unit test for run_z_probe development
Simple make file, ignore resulting test executable.

Rework adaptive probe to search through all the returned results

This avoids the problem the previous could run into, which was if 2 of
the first 3 probes are both bad, but closer to each other than all the
following good values, each of the good values would be thrown away as
worse than the two bad values.

Redo adaptive probe algorithm, pass if good - bad > 2

This avoids the problem where if two of the bad values agree more than
individual bad values and the good value, it will discard the good
value and try again until it runs out of retries.  This runs an
evaluation after each new point to determine good or bad and if it has
enough good values to proceed, then averages the good values.