Redo adaptive probe algorithm, pass if good - bad > 2
This avoids the problem where if two of the bad values agree more than
individual bad values and the good value, it will discard the good
value and try again until it runs out of retries.  This runs an
evaluation after each new point to determine good or bad and if it has
enough good values to proceed, then averages the good values.