G28 H Z, home Z at the current X/Y independent of POWER_LOSS_RECOVERY
"Fix, Improve Power-Loss Recovery (#22828)" enabled this H option, but
only when POWER_LOSS_RECOVERY is defined.  This is useful beyond power
loss recover, such as printing from OctoPrint, or any time there may
be a model at the normal safe x/y location and a Z home is needed.
It is also more flexible in letting the user choose the location based
on where the model was sliced for this print.  Add comments to
document H.

In my latest recovery I first needed the model height, home XY, home Z
(avoiding the model), move to the model, G30 to probe the model
height.  Then I could restart the print at that location.  Printing
from OctoPrint power loss recovery was not involved.

Note the probe to nozzle offset means some locations next to the edge
of the build plate will probe once, then fail because it sees the
location as unreachable and fail the home Z.