G30 select relative or absolute positions
Absolute is where to move the print head (nozzle), relative will move
the probe to that bed location using the probe offset.  Without X or Y
this will now use absolute so it will probe straight down  instead of
first moving the probe to that location.  This is useful when jogging
the probe to a model you want to measure, or just repeating the
measurement at the current location compared to always using relative
tool head would move over with each measurement.  From the G30
documentation "By default probe in the current position."  Which I
take to be straight down, otherwise it requires