BLTouch::command Avoid write for duplicate probe commands
deploy(); deploy(); or the reverse stow(); stow(); would delay by
default 750 ms each command giving time to check for an alarm.  If the
command is the same as the last command, it would have already delayed
checking for an alarm, skip the write, and skip the delay.

Here are the calls that result in the duplicate command at the start
and end of a set or probes.  Handling it on the BLTouch side seems
better than coming up with a solution to try to avoid the duplicate
calls in the nested functions.

Probe::probe_at_point() calls deploy(), run_z_probe(), stow()
run_z_probe() calls probe_down_to_z()
probe_down_to_z() calls bltouch.deploy(), bltouch.stow()