Move bltouch._reset() after the move to a safe height
If the probe is currently being triggered at the start of
probe_at_point(), it could be currently touching something or it could
be alarming.  This is about to move up to a safe z height anyway,
which could clear the trigger condition.  The move will also give it
additional time to settle if it is transient.  If it is still
triggering after the move it might be the alarm, so do the reset.
Once it is reset, the probe needs to be deployed, and there is already
logic to handle that.

Remove only checking if it is in high_speed_mode, this catches the
alarm condition either way.