Decrease the total time by moving Z at the same time as XY
I was thinking about how I could make auto bed leveling even faster,
consider this one for fun.  This is a compromise between speed and
safety more on the speed side.

This raises Z until the probe clears the bed (and whatever risk
tolerance the user has), then takes a straight line diagonal XYZ to
the probe point.  XY are likely moving further, Z is normally slower,
don't let Z slow down XY, which likely gives a little more initial Z
raise.

This requires G29 to not ask it to raise after the probe point, but
rather wait for the following movement, so it has both the Z travel
and XY movement.