probe_down_to_z skip passing constants as arguments
The lambda function can use the local variables directly.
This changes the early_fail check, my Z offset is -2.52 and for
BLTouch high speed probe 2nd and on probe, I'm having success with
a raise of only 1mm, which means Z_CLEARANCE_MULTI_PROBE is -1.52
which was causing the early trigger failure.